requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
requests
yfinance
yahooquery
pyarrow
python-dotenv
lxml
html5lib
beautifulsoup4
tqdm
pymongo
Flask
